requirements.txt
asgiref==3.8.1
Django==5.2.1
django-widget-tweaks==1.5.0
django-debug-toolbar==4.4.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
